Update sample_output_octant_analysis_mod_5000.xlsx
A very pertinent query asked on the Google form by 2001ME32
https://docs.google.com/spreadsheets/d/1QiKySHoGXoG8hOUhG7saDMI7eV4yOqK2LCTY2bt-Lys/edit#gid=2139008478

"When we are calculating transition values for mod, then we need to take mod ranges like 0 to 4999, then 4999 to 9998,... And so on. If take mod ranges from 0 to 4999 , 5000 to 9999... And so on, then we shall miss the transition of 4999 to 5000. We shall traverse our octant values from 0 to 4999, then from 5000 to 9999 so the transition from 4999 to 5000 shall be missed ."

So, in case of transition computation for tut02, it will be:

0 to 4999
4999 to 9998
9998 to 14997
14997 to 19996
19996 to 24995
24995 to 29994

So, in each band the elements shall be 5000 elements

Thanks 2001ME32 for the wonderful observation.